Replace the gut of shellshim mock
